+++ backend/canisters/user_index/impl/src/updates/c2c_moderation_referral.rs
@@ -69,6 +69,7 @@ fn c2c_moderation_referral_impl(args: Args, state: &mut RuntimeState) {
             blob_references: args.blob_references.clone(),
             media_matches: Vec::new(),
             timestamp: now,
+            authority_report: None,
         },
         state,
     );